improve the handling of exceptional cases
When pruning is avoided (because no results match), print a
message explaining this at the top of the page. Otherwise this
'non-monotone' behavior might be confusing to users.

Also, say you are reading the etymology of some word, and you click
on a link to a word in a language that would normally be pruned,
adding this language to the URL like '#Greek'. Assume furthermore
that this word does exist in some other non-pruned language. Then,
at this page, 'Greek' should not be pruned. Otherwise the link
would be malfunctioning.